feat: 일기 데이터 추가 및 편집 기능 구현

+++ Diary/Diary+CoreDataProperties.swift
@@ -17,7 +17,7 @@ extension Diary {
 
     @NSManaged public var title: String?
     @NSManaged public var body: String?
-    @NSManaged public var createdAt: Int64
+    @NSManaged public var createdAt: Double
     @NSManaged public var id: UUID?
 
 }
